Add article enrichment pipeline
